Authorization bearer

+++ api/Controllers/MoviesController.js
@@ -5,6 +5,9 @@ class MoviesController {
  
     async createMovie(req, res) {
       try {
+			const token = securityService.getTokenFromHeader(req);
+			if (token == null || token == undefined) return res.status(401).json({ error: 'Token no proporcionado' });
+			if(securityService.verifyToken(token) != true) return res.status(403).json({error: 'token no validado'});
         const Movie = await moviesService.createMovie(req.body);
         res.status(201).json(Movie);
       } catch (err) {
@@ -34,6 +37,9 @@ class MoviesController {
 
     async UpdateMovie(req, res){
         try{
+			const token = securityService.getTokenFromHeader(req);
+			if (token == null || token == undefined) return res.status(401).json({ error: 'Token no proporcionado' });
+			if(securityService.verifyToken(token) != true) return res.status(403).json({error: 'token no validado'});
             const Movies = await moviesService.updateMovie(req.params.id, req.body);
             if(!Movies) res.status(404).json({error: 'not found'});
             res.status(201).json(Movies);
@@ -44,6 +50,9 @@ class MoviesController {
 
     async DeleteMovie(req, res){
         try{
+			const token = securityService.getTokenFromHeader(req);
+			if (token == null || token == undefined) return res.status(401).json({ error: 'Token no proporcionado' });
+			if(securityService.verifyToken(token) != true) return res.status(403).json({error: 'token no validado'});
             const Movies = await moviesService.deleteMovie(req.params.id);
             if(!Movies) res.status(404).json({error: 'not found'});
             res.status(201).json(Movies);
