+++ api/Controllers/SalasController.js
@@ -6,6 +6,9 @@ class SalasController {
  
     async createSala(req, res) {
       try {
+			const token = securityService.getTokenFromHeader(req);
+			if (token == null || token == undefined) return res.status(401).json({ error: 'Token no proporcionado' });
+			if(securityService.verifyToken(token) != true) return res.status(403).json({error: 'token no validado'});
         const Sala = await salasService.createSala(req.body);
         res.status(201).json(Sala);
       } catch (err) {
@@ -34,6 +37,9 @@ class SalasController {
 
     async UpdateSala(req, res){
         try{
+			const token = securityService.getTokenFromHeader(req);
+			if (token == null || token == undefined) return res.status(401).json({ error: 'Token no proporcionado' });
+			if(securityService.verifyToken(token) != true) return res.status(403).json({error: 'token no validado'});
             const Salas = await salasService.updateSala(req.params.id, req.body);
             if(!Salas) res.status(404).json({error: 'not found'});
             res.status(201).json(Salas);
@@ -44,6 +50,9 @@ class SalasController {
 
     async DeleteSala(req, res){
         try{
+			const token = securityService.getTokenFromHeader(req);
+			if (token == null || token == undefined) return res.status(401).json({ error: 'Token no proporcionado' });
+			if(securityService.verifyToken(token) != true) return res.status(403).json({error: 'token no validado'});
             const Salas = await salasService.deleteSala(req.params.id);
             if(!Salas) res.status(404).json({error: 'not found'});
             res.status(201).json(Salas);
@@ -51,7 +60,7 @@ class SalasController {
             res.status(500).json({error: err.message});
         }
     }
-
+//cambiar en un futuro
     async ModificarPeliculaSala(req, res){
         try{
             const Salas = await salasService.ModificarPeliculaSala(req.params.id,req.params.idMovie);
@@ -61,7 +70,7 @@ class SalasController {
             res.status(500).json({error: err.message});
         }
     }
-
+//cambiar en un futuro
     async ModificarAsiento(req, res){
         try{
             const Salas = await salasService.ModificarAsiento(req.params.id,req.body);
