+++ api/Controllers/UsuariosController.js
@@ -1,5 +1,7 @@
 import UsuariosService from "../services/UsuariosService.js";
+import SecurityService from "../services/SecurityService.js";
 const usuariosService = new UsuariosService;
+const securityService = new SecurityService;
 class UsuariosController {
  
  
@@ -13,7 +15,11 @@ class UsuariosController {
     }
 
     async getUsuario(req, res){
-        try{
+		try{
+			const token = securityService.getTokenFromHeader(req);
+			if (token == null || token == undefined) return res.status(401).json({ error: 'Token no proporcionado' });
+			if(securityService.verifyToken(token) != true) return res.status(403).json({error: 'token no validado'});
+
             const Usuario = await usuariosService.getUsuario(req.params.id);
             if(!Usuario) res.status(404).json({error: 'not found'});
             res.status(200).json(Usuario);
@@ -24,6 +30,10 @@ class UsuariosController {
 
 	 async getUsuarioByUsername(req, res){
         try{
+			const token = securityService.getTokenFromHeader(req);
+			if (token == null || token == undefined) return res.status(401).json({ error: 'Token no proporcionado' });
+			if(securityService.verifyToken(token) != true) return res.status(403).json({error: 'token no validado'});
+
             const Usuario = await usuariosService.getUsuarioByUsername(req.params.username);
             if(!Usuario) res.status(404).json({error: 'not found'});
             res.status(200).json(Usuario);
@@ -35,6 +45,10 @@ class UsuariosController {
 
     async UpdateUsuario(req, res){
         try{
+			const token = securityService.getTokenFromHeader(req);
+			if (token == null || token == undefined) return res.status(401).json({ error: 'Token no proporcionado' });
+			if(securityService.verifyToken(token) != true) return res.status(403).json({error: 'token no validado'});
+
             const Usuario = await usuariosService.updateUsuario(req.params.id, req.body);
             if(!Usuario) res.status(404).json({error: 'not found'});
             res.status(200).json(Usuario);
@@ -45,6 +59,10 @@ class UsuariosController {
 
     async DeleteUsuario(req, res){
         try{
+			const token = securityService.getTokenFromHeader(req);
+			if (token == null || token == undefined) return res.status(401).json({ error: 'Token no proporcionado' });
+			if(securityService.verifyToken(token) != true) return res.status(403).json({error: 'token no validado'});
+
             const Usuario = await usuariosService.deleteUsuario(req.params.id);
             if(!Usuario) res.status(404).json({error: 'not found'});
             res.status(200).json(Usuario);
